ADD Camera Class

+++ includes/game/Settings.hpp
@@ -2,8 +2,8 @@
 
 // constexpr int WIDTH = 1920;
 // constexpr int HEIGHT = 1080;
-constexpr GLuint		WIDTH = 1080;
-constexpr GLuint		HEIGHT = 720;
+constexpr GLfloat		WIDTH = 1080.0;
+constexpr GLfloat		HEIGHT = 720.0;
 // constexpr int WIDTH = 800;
 // constexpr int HEIGHT = 800;
 constexpr const char	*TITLE = "GAME";
